Extensive rework of the palette code.
The Median Cut algorithm is now pluggable,
and 2 implementations have been added, the old being
the longest axis cutting and the new is similar to
the one used by the PaletteFactory's own version of
Median Cut (which seems to give much better results).

static final constants in MedianCutPalette have been
converted to an enum.

Various cleanups.



git-svn-id: https://svn.apache.org/repos/asf/commons/proper/imaging/trunk@1420371 13f79535-47bb-0310-9956-ffa450edef68