Extensive refactoring and cleanup.
Encapsulated, documented and simplified TiffField variables, and
made only a single value variable instead of both an offset and value.
Initial work on use of long for offsets, so that files
> 2 GB can be read. Cleanup of excessive
interface inheritance that was used to pull in constants.
Generification of the LZW compressor.

The API and ABI will come apart at the seams :-/, but in the long run,
it's for the better.



git-svn-id: https://svn.apache.org/repos/asf/commons/proper/imaging/trunk@1445822 13f79535-47bb-0310-9956-ffa450edef68